py-Backend/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
dlib==19.24.6
dnspython==2.6.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
PyJWT==2.9.0
pymongo==4.10.1
pytz==2024.2
six==1.16.0
Werkzeug==3.0.6
zipp==3.20.2